Add workflows to push to GHCR (#7)
* testing pushing to ghcr

* Added workflows for pushing to GHCR

* Added auth

* Use powershell core

* Allow the repo to not exist

* Fixed workflow names

* Use the owner variable

* Fixed images names

* Fixed workflow name

* Deal with an empty errors array

* Fixed logic for not continuing

* The repo owner needs to be lowercase

* Make version comparison field consistent to remove image specific one

Renamed $latestFlywayVersion to $versionToCompare as thats consistent with other builds. There was a small error in the GHCR version so streamlining this might avoid a copy-paste error in the future

* Make version comparison field consistent to remove image specific one

Renamed $latestLiquibaseVersion to $versionToCompare (where necessary) as thats consistent with other builds. There was a small error in the GHCR version so streamlining this might avoid a copy-paste error in the future

---------

Co-authored-by: Mark Harrison <mark.harrison@octopus.com>